+++ src/ts/Ads/Utilities/SDKMetrics.ts
@@ -152,6 +152,7 @@ export enum AUIMetric {
     InvalidImageAssets = 'invalid_image_assets',
     InvalidCtaText = 'invalid_cta_text',
     InvalidSchemeAndColorCoordination = 'invalid_scheme_and_color_coordination',
+    ColorMatchingNotSupported = 'color_matching_not_supported',
     TiltedLayoutNotSupported = 'tilted_layout_not_supported'
 }